requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
flake8
python-dotenv>=0.5.1

setuptools==69.0.3
sqlalchemy==2.0.25
pandas==2.1.4
pymysql==1.1.0
mysql-connector-python==8.2.0
pyarrow==14.0.2
fastparquet==2023.10.1
werk24==1.7.0
phpserialize==1.3
jax==0.4.23
libclang==16.0.6
tensorflow==2.15.0
scikit-learn==1.3.2
matplotlib==3.8.0
seaborn==0.13.1
keras==2.15.0
jupyterlab==4.0.8
notebook==7.0.6
jupyter-nbextensions-configurator==0.6.3